pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "discord-raw-mcp"
version = "0.1.0"
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
description = "Discord Raw API MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.2",
    "mcp>=1.2.0",
    "aiohttp>=3.9.1"
]

[project.scripts]
discord-raw-mcp = "discord_raw_mcp:main"